Rename two classes.
`ConfigFile` was a compromise to begin with, to avoid
name clashes with hiera's own config class. With the new
namespace this is no longer a concern.

`ReadFile` was confusing, as it is also used to write and
remove files. I opted for `YamlFile` because this is ultimately
what this class (currently) represents. This might change if we
add support for e.g. JSON.