Remove encrypted credentials file #56
Add instructions how to generate a new one.

Also add the file to .gitignore. This is not standard
practice for rails applications in general, but makes sense
for an open source application that will have a lot of
different installations. Each of those should have their own
`secrect_key_base` (and possibly other credentials that
might be added in the future).