Fix and improve node environment handling #111
This fixes a few of the WIP changes to consistently
provide and use an environment when querying for nodes.

The `Node` model now no longer represents a unique
hostname, but a unique combination of hostname and
enviroment.

Also removes the "validations" in environment and node
constructors. These were kind of redundant and also prevented
using these models to represent new, not-yet-persisted records.