Large refactoring #4
With the old funtionality it was OK to always need a node,
because a user had to choose one after choosing an environment.

With #4 we are introducing a different funtionality that
should work indepently of a node.

This is in preparation of #4 and tries to disentangle
environment, nodes and everything else.

To accomplish this, "proper" abstractions for all relevant
parts of the hiera data model are introduced (hierarchies,
data_files and values).

Due to the git backend, (data_)file and value handling still
need to be able to work with a node's facts, but this has
been made optional.