HDM API (#153)
* Add role to users #87

Role can be either `admin`, `regular` or `api`. The first
two represent the two existing roles while the third is
added in preparation for API access.

* First steps towards an API #87

Allows to query for environments, nodes and keys for now.

Returned JSON structure might not yet be final.

* Allow querying key data w/o environment #87

This will simply use the node's environment.