Migration to Bootstrap 5 (#216)
* Migration to Bootstrap 5.

Since Bs5 no longer depends on jQuery, and we do not use it
anywhere else, this also removes that dependency.

Some visual changes were unavoidable, but overall I took
care not to change too much at once and existing users should
feel right at home.

* Do not precompile jquery...

... as the dependency on jquery was removed.

* Fix button text color.

With the old BS4 code, the primary color was applied
automatically in this case. With the new accordion
component in BS5 we need to add it explicitly.