Improve first access flow #234
Adds a welcome page upon first access that explains that
an admin user is needed and displays information about
external authentication sources if applicable.

It no longer prompts to create a new regular user if
external authentication is already configured.

If authentication is disabled none of this is displayed
(fixes #54)