Add diff view to lookup results #301
Just like with regular value display from a single file this
now checks if a node is from a different environment than
the selected one.

If it is, the same tabbed view with a diff is presented,
showing the lookup result in the selected environment and
the diff to the lookup in the original environment.