Add configurable node info section to page headers
In `config/hdm.yml` set

```yaml
node_info:
  - facts.name1
  - facts.name2
  - # ...
```

and include up to 6 fact names. These names and their
respective values will be shown at the top of a node's
pages to help provide more context.

The info box is hidden if `node_info` is not set.

This also adds a headline with the node's name to
all node pages irrespective of whether `node_info` is
set or not.

Signed-off-by: David Roetzel <david@roetzel.de>