Enhance ldap configuration (#588, #593) (#601)
This allows setting a different username attribute (e.g. `uid`
instead of the default, `mail`).

It also allows specifying an additional filter that is added
when trying to authenticate a user. This can be used to e.g. limit
ldap authentication to a group (but really anything).

Last but not least, TLS options have been enhanced. "STARTTLS"
should now be possible, also a CA file can be specified.
Certificate verification has been turned on as a new default,
but can optionally be turned off.

Co-authored-by: Martin Alfke <tuxmea@users.noreply.github.com>